Gravity force applies to particle
